refactor(test): merge function and unit tests for template strings
To avoid confusing name ending with `e2e` in the unit tests package.
