--- core/test/unit/src/template-string/template-strings-e2e.ts
@@ -1,28 +1,0 @@
-/*
- * Copyright (C) 2018-2023 Garden Technologies, Inc. <info@garden.io>
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-
-import { getDataDir, makeTestGarden, TestGarden } from "../../../helpers"
-import { expect } from "chai"
-
-describe("template-strings", () => {
-  context("cross-context variable references", () => {
-    let dataDir: string
-    let garden: TestGarden
-
-    before(async () => {
-      dataDir = getDataDir("test-projects", "template-strings")
-      garden = await makeTestGarden(dataDir)
-    })
-
-    it("should resolve variables from project-level and environment-level configs", async () => {
-      const graph = await garden.getConfigGraph({ log: garden.log, emit: false })
-      const deployAction = graph.getDeploy("test-deploy")
-      expect(deployAction.getConfig().include).to.eql(["aFileFromEnvConfig", "aFileFromProjectConfig"])
-    })
-  })
-})
